refactor(任务中心): 优化任务中心按照任务类型加载
