--- api-test/frontend/src/business/history/task/ApiTaskPagination.vue
@@ -1,62 +1,0 @@
-<template>
-  <el-row type="flex" justify="end">
-    <div class="table-page">
-      <el-pagination
-        background
-        @size-change="handleSizeChange"
-        @current-change="handleCurrentChange"
-        :current-page="currentPage"
-        :page-sizes="pageSizes"
-        :page-size="pageSize"
-        :page-count="5"
-        small
-        layout="total, sizes, prev, pager, next"
-        :total="total">
-      </el-pagination>
-    </div>
-  </el-row>
-</template>
-
-<script>
-  export default {
-    name: "MsTaskPagination",
-    props: {
-      page: Object,
-      currentPage: {
-        type: Number,
-        default: 1
-      },
-      pageSize: {
-        type: Number,
-        default: 5
-      },
-      pageSizes: {
-        type: Array,
-        default: function () {
-          return [5, 10, 20, 50, 100]
-        }
-      },
-      total: {
-        type: Number,
-        default: 0
-      },
-      change: Function
-    },
-    methods: {
-      handleSizeChange: function (size) {
-        this.$emit('update:pageSize', size)
-        this.change();
-      },
-      handleCurrentChange(current) {
-        this.$emit('update:currentPage', current)
-        this.change();
-      }
-    }
-  }
-</script>
-
-<style scoped>
-  .table-page {
-    padding-top: 10px;
-  }
-</style>
