+++ framework/sdk-parent/sdk/src/main/java/io/metersphere/base/mapper/ext/BaseTaskMapper.java
@@ -8,7 +8,10 @@ import java.util.List;
 
 public interface BaseTaskMapper {
 
-    List<TaskCenterDTO> getTasks(@Param("request") TaskCenterRequest request, @Param("uiPermission") Boolean uiPermission);
+    List<TaskCenterDTO> getApiTasks(@Param("request") TaskCenterRequest request);
+    List<TaskCenterDTO> getScenarioTasks(@Param("request") TaskCenterRequest request);
+    List<TaskCenterDTO> getPerfTasks(@Param("request") TaskCenterRequest request);
+    List<TaskCenterDTO> getUiTasks(@Param("request") TaskCenterRequest request, @Param("uiPermission") Boolean uiPermission);
 
     int getRunningTasks(@Param("request") TaskCenterRequest request);
