+++ framework/sdk-parent/sdk/src/main/java/io/metersphere/commons/constants/TaskCenterType.java
@@ -1,0 +1,5 @@
+package io.metersphere.commons.constants;
+
+public enum TaskCenterType {
+    API, SCENARIO, UI, PERF
+}
