+++ framework/sdk-parent/sdk/src/main/java/io/metersphere/task/dto/TaskCenterRequest.java
@@ -28,5 +28,7 @@ public class TaskCenterRequest {
 
     private int pageSize;
 
+    private String activeName;
+
     List<String> projects;
 }
