+++ framework/sdk-parent/sdk/src/main/java/io/metersphere/task/service/TaskService.java
@@ -1,9 +1,14 @@
 package io.metersphere.task.service;
 
+import com.github.pagehelper.Page;
+import com.github.pagehelper.PageHelper;
 import io.metersphere.base.mapper.ext.BaseInformationSchemaTableMapper;
 import io.metersphere.base.mapper.ext.BaseScheduleMapper;
 import io.metersphere.base.mapper.ext.BaseTaskMapper;
 import io.metersphere.commons.constants.MicroServiceName;
+import io.metersphere.commons.constants.TaskCenterType;
+import io.metersphere.commons.utils.PageUtils;
+import io.metersphere.commons.utils.Pager;
 import io.metersphere.dto.TaskInfoResult;
 import io.metersphere.request.BaseQueryRequest;
 import io.metersphere.service.BaseCheckPermissionService;
@@ -56,11 +61,45 @@ public class TaskService {
         return new ArrayList<>(userRelatedProjectIds);
     }
 
-    public List<TaskCenterDTO> getTasks(TaskCenterRequest request) {
+    public Pager<List<TaskCenterDTO>> getTasks(TaskCenterRequest request) {
+        Page<Object> page = PageHelper.startPage(request.getGoPage(), request.getPageSize(), true);
+        if (StringUtils.equals(request.getActiveName(), TaskCenterType.API.name())) {
+            return PageUtils.setPageInfo(page, getApiTasks(request));
+        } else if (StringUtils.equals(request.getActiveName(), TaskCenterType.SCENARIO.name())) {
+            return PageUtils.setPageInfo(page, getScenarioTasks(request));
+        } else if (StringUtils.equals(request.getActiveName(), TaskCenterType.PERF.name())) {
+            return PageUtils.setPageInfo(page, getPerfTasks(request));
+        } else {
+            return PageUtils.setPageInfo(page, getUiTasks(request));
+        }
+    }
+
+    public List<TaskCenterDTO> getApiTasks(TaskCenterRequest request) {
         if (CollectionUtils.isEmpty(request.getProjects())) {
             return new ArrayList<>();
         }
-        return baseTaskMapper.getTasks(request,checkUiPermission());
+        return baseTaskMapper.getApiTasks(request);
+    }
+
+    public List<TaskCenterDTO> getUiTasks(TaskCenterRequest request) {
+        if (CollectionUtils.isEmpty(request.getProjects())) {
+            return new ArrayList<>();
+        }
+        return baseTaskMapper.getUiTasks(request, checkUiPermission());
+    }
+
+    public List<TaskCenterDTO> getPerfTasks(TaskCenterRequest request) {
+        if (CollectionUtils.isEmpty(request.getProjects())) {
+            return new ArrayList<>();
+        }
+        return baseTaskMapper.getPerfTasks(request);
+    }
+
+    public List<TaskCenterDTO> getScenarioTasks(TaskCenterRequest request) {
+        if (CollectionUtils.isEmpty(request.getProjects())) {
+            return new ArrayList<>();
+        }
+        return baseTaskMapper.getScenarioTasks(request);
     }
 
     public int getRunningTasks(TaskCenterRequest request) {
@@ -95,7 +134,7 @@ public class TaskService {
             if (taskRequestMap.containsKey(PERF)) {
                 microService.postForData(MicroServiceName.PERFORMANCE_TEST, "/performance/stop/batch", taskRequestMap.get(PERF));
             }
-            if(taskRequestMap.containsKey(UI)){
+            if (taskRequestMap.containsKey(UI)) {
                 microService.postForData(MicroServiceName.UI_TEST, "/ui/automation/stop/batch", reportIds);
             }
         }
@@ -115,13 +154,13 @@ public class TaskService {
         }
     }
 
-    private boolean checkUiPermission(){
+    private boolean checkUiPermission() {
         try {
             String uiScenarioReport = baseInformationSchemaTableMapper.checkExist(UI_SCENARIO_REPORT);
-            if(StringUtils.isNotEmpty(uiScenarioReport)){
+            if (StringUtils.isNotEmpty(uiScenarioReport)) {
                 return true;
             }
-        }catch (Exception e){
+        } catch (Exception e) {
         }
         return false;
     }
